Upgrade to Version 2.1
Version 2.1 reworks the search system, introducing a set of search
parameters that can be used to control and restrict output for a query.
These search parameters take the form of grades or homeforms that can
be manually excluded from the output of a search. For instance, it
might be of interest to list only the grade 12's in a group of people,
of even as specific as the grade 12's in homeform 12A. With search
parameters, a user can exclude all grades except grade 12 and all
homeforms except 12A to have this effect.

New class SearchParametersFrame provides a convenient interface for
setting or unsetting these search parameters, including several utility
buttons to select or clear a group of buttons at once, or to revert to
defaults. An instance of this frame can be accessed by pressing a new
button at the bottom of the main frame, labelled 'Set Parameters'.

In addition to the Set Parameters button, a new button has been added
to remove all text from all input and output text panes, labelled
'Clear Text Frames'. The system of search buttons has been reworked,
and instead of three separate buttons to launch searches for single
pieces of information at a time, several new toggle buttons allow the
user to choose which pieces of information to report, possibly more
than one at a time. The search is then launched by pressing a single
button labelled 'Search'.

A search can now be placed on the whole contents of the index file,
instead of on a provided list of names, by clearing the input text
pane for the multiple-row input space and launching a search.
Any search parameters still apply.

These changes required updates in the structure of several other
classes, including data classes like PeopleDataList, RequestEvent, and
IndexInterpreter. In addition some bugs were fixed, including an
off-by-one error that would cause the first person in every index to
always be reported as spelled wrong, and a bug that prevented scroll
bars from properly matching and unmatching when text was edited after
a search.