Remember old query strings
Instead of querying the exact text within the input frame, the mugs
reader now only updates the query string when the user modifies the
input data. This means different parameters can be used on a single
query string, switching grades, homeforms, etc, without needing to
re-enter the query string every time.

For example, if the whole yearbook team is used as a query string
to determine members of each grade for distribution, then grade 10
and grade 11 can be searched consecutively, and both will search the
whole yearbook team instead of the second query searching the results
of the first.